Use `ComparableVersion` instead of custom impl
